login/register functionality updated for both new and exists user

+++ pages/Almost There Page After NumberSignUp/AlmostThere.js
@@ -13,13 +13,28 @@ SAVE.addEventListener("click", function () {
         alert("fields can't be empty")
         return;
     }
-    sessionStorageData["Name"] = fullName.value;
-    sessionStorageData["email"] = email.value;
+    if (!email.value.length >= 10) {
+        alert("enter a valid email minimum length: 10")
+        return;
+    }
 
-    sessionStorage.setItem("userDetails", JSON.stringify(sessionStorageData));
-    alert(` ${fullName.value} Successfully registered`);
-    window.location.href = "/index.html"
-
+    //saving registered user into localstorage 
+    let AllRegisteredUsers = JSON.parse(localStorage.getItem("AllRegisteredUsers")) || []
+    // checking if the email already exists
+    let isEmailExists = AllRegisteredUsers.some((obj) => obj.email == email.value)
+    if (isEmailExists) {
+        alert("email exists user a different one or login")
+        return;
+    } else {
+        sessionStorageData["Name"] = fullName.value;
+        sessionStorageData["email"] = email.value;
+        AllRegisteredUsers.push(sessionStorageData)
+        localStorage.setItem("AllRegisteredUsers", JSON.stringify(AllRegisteredUsers));
+        // and a copy session storage for logout purpose
+        sessionStorage.setItem("userDetails", JSON.stringify(sessionStorageData));
+        alert(` ${fullName.value} Successfully registered`);
+        window.location.href = "/index.html"
+    }
 
 
 })
